按钮显示限制，百分数显示

+++ Assets/Scripts/PartnerData.cs
@@ -25,7 +25,7 @@ public class PartnerData : MonoBehaviour {
         }
 
         Text_RecentRoomCount.text = "近10个牌局两人同桌 <color=#18ffff>" + data.Int("recent_room_count") + "</color> 局";
-        Text_RecentRoomProfit.text = "近10个同桌牌局盈利率为： <color=#ffca28>" + data.Int("recent_room_profit") * 100 + "%</color>";
-        Text_TotalRoomProfit.text = "所有同桌牌局总盈利率为： <color=#ffca28>" + data.Int("total_room_profit") * 100 + "%</color>";
+        Text_RecentRoomProfit.text = "近10个同桌牌局盈利率为： <color=#ffca28>" + data.Float("recent_room_profit") * 100 + "%</color>";
+        Text_TotalRoomProfit.text = "所有同桌牌局总盈利率为： <color=#ffca28>" + data.Float("total_room_profit") * 100 + "%</color>";
     }
 }
